refactor: batch 0 part 1 – smazání broken BMI a duplicitních -new routes
- Smazán broken BMI v1 (undefined translations)
- Obnověna /bmi route z bmi-new (bmi-new smazán)
- Smazány všechny -new placeholdery (anuita, cista-mzda, kolik-procent, y-je-x, prevodnik, slozene, trojclenka, zlomky)
- Smazány dph-new a procento-z-cisla-new (původní verze jsou bohatší)
- Přidán redirect /bmi → /bmi-new do middleware.ts

Výsledek: 0 duplicitních route pairs
