--- src/app/[locale]/calculator/anuita-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/anuita-new/page.tsx
-'use client';
-
-import AnnuityPaymentCalculator from '@/components/calculators/AnnuityPaymentCalculator';
-
-export default function AnnuityPaymentTestPage() {
-  return <AnnuityPaymentCalculator />;
-}
