+++ src/app/[locale]/calculator/bmi/page.tsx
@@ -1,29 +1,24 @@
 'use client';
 
-import React, { Suspense, useMemo } from 'react';
+import React, { Suspense } from 'react';
 import dynamic from 'next/dynamic';
-import { useTranslations } from 'next-intl';
 import { Skeleton } from '@/components/ui/skeleton';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/Card';
-import { Info } from 'lucide-react';
-import { 
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip';
 
-// Dynamically import components with loading states
-const CalculatorComponent = dynamic(() => import('@/components/calculators/BMICalculator'), {
-  loading: () => (
-    <div className="space-y-4">
-      <Skeleton className="h-10 w-1/2" />
-      <Skeleton className="h-8 w-full" />
-      <Skeleton className="h-40 w-full" />
-    </div>
-  ),
-  ssr: false
-});
+// Dynamically import the BMI calculator
+const BMICalculator = dynamic(
+  () => import('@/components/calculators/BMICalculator'),
+  {
+    loading: () => (
+      <div className="space-y-4 p-6">
+        <Skeleton className="h-8 w-1/2" />
+        <Skeleton className="h-4 w-3/4" />
+        <Skeleton className="h-40 w-full" />
+        <Skeleton className="h-32 w-full" />
+      </div>
+    ),
+    ssr: false
+  }
+);
 
 // Error boundary for calculator component
 interface ErrorBoundaryProps {
@@ -42,7 +37,7 @@ class ErrorBoundary extends React.Component<ErrorBoundaryProps, { hasError: bool
   }
 
   componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
-    console.error('Calculator error:', error, errorInfo);
+    console.error('BMI Calculator error:', error, errorInfo);
   }
 
   render() {
@@ -53,41 +48,35 @@ class ErrorBoundary extends React.Component<ErrorBoundaryProps, { hasError: bool
   }
 }
 
-const CalculatorPage = () => {
-  const t = useTranslations();
-
-  // Memoize translations to prevent unnecessary re-renders
-  const translations = useMemo(() => ({
-    seoTitle: t('bmi_title') || 'Bmi',
-    seoDescription: t('bmi_seo_description') || 'Calculator description',
-    tip: t('bmi_tip') || 'Enter values to calculate',
-    tipText: t('bmi_tip_text') || 'Enter values to see the calculation',
-    loadingError: t('chyba_nacitani') || 'Failed to load calculator. Please try refreshing the page.'
-  }), [t]);
-
-  const { seoTitle, seoDescription, tip, tipText, loadingError } = translations;
-
-  // Translations are loaded via useTranslations hook
-
+const BMINewPage: React.FC = () => {
   return (
-    <ErrorBoundary 
+    <ErrorBoundary
       fallback={
-        <div className="text-destructive p-4 rounded-lg bg-destructive/10">
-          <p>{loadingError}</p>
+        <div className="container mx-auto px-4 py-8">
+          <div className="text-center">
+            <h1 className="text-2xl font-bold text-red-600 mb-4">
+              Chyba při načítání kalkulátoru
+            </h1>
+            <p className="text-gray-600">
+              Omlouváme se, došlo k chybě při načítání BMI kalkulátoru. 
+              Zkuste prosím obnovit stránku.
+            </p>
+          </div>
         </div>
       }
     >
       <Suspense fallback={
-        <div className="space-y-4">
-          <Skeleton className="h-10 w-1/2" />
-          <Skeleton className="h-8 w-full" />
+        <div className="space-y-4 p-6">
+          <Skeleton className="h-8 w-1/2" />
+          <Skeleton className="h-4 w-3/4" />
           <Skeleton className="h-40 w-full" />
+          <Skeleton className="h-32 w-full" />
         </div>
       }>
-        <CalculatorComponent />
+        <BMICalculator />
       </Suspense>
     </ErrorBoundary>
   );
 };
 
-export default CalculatorPage;
+export default BMINewPage;
