--- src/app/[locale]/calculator/cista-mzda-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/cista-mzda-new/page.tsx
-'use client';
-
-import NetSalaryCalculator from '@/components/calculators/NetSalaryCalculator';
-
-export default function NetSalaryTestPage() {
-  return <NetSalaryCalculator />;
-}
