--- src/app/[locale]/calculator/kolik-procent-je-x-z-y-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/kolik-procent-je-x-z-y-new/page.tsx
-'use client';
-
-import WhatPercentageIsXOfYCalculator from '@/components/calculators/WhatPercentageIsXOfYCalculator';
-
-export default function WhatPercentageTestPage() {
-  return <WhatPercentageIsXOfYCalculator />;
-}
