--- src/app/[locale]/calculator/prevodnik-jednotek-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/prevodnik-jednotek-new/page.tsx
-'use client';
-
-import UnitConverter from '@/components/calculators/UnitConverter';
-
-export default function UnitConverterTestPage() {
-  return <UnitConverter />;
-}
