--- src/app/[locale]/calculator/procento-z-cisla-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/procento-z-cisla-new/page.tsx
-'use client';
-
-import PercentageOfNumberCalculator from '@/components/calculators/PercentageOfNumberCalculator';
-
-export default function PercentageOfNumberTestPage() {
-  return <PercentageOfNumberCalculator />;
-}
