--- src/app/[locale]/calculator/slozene-uroceni-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/slozene-uroceni-new/page.tsx
-'use client';
-
-import CompoundInterestCalculator from '@/components/calculators/CompoundInterestCalculator';
-
-export default function CompoundInterestTestPage() {
-  return <CompoundInterestCalculator />;
-}
