--- src/app/[locale]/calculator/trojclenka-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/trojclenka-new/page.tsx
-'use client';
-
-import DirectProportionCalculator from '@/components/calculators/DirectProportionCalculator';
-
-export default function RuleOfThreeTestPage() {
-  return <DirectProportionCalculator />;
-}
