--- src/app/[locale]/calculator/y-je-x-kolik-je-sto-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/y-je-x-kolik-je-sto-new/page.tsx
-'use client';
-
-import YIsXWhatIsHundredCalculator from '@/components/calculators/YIsXWhatIsHundredCalculator';
-
-export default function YIsXPercentTestPage() {
-  return <YIsXWhatIsHundredCalculator />;
-}
