--- src/app/[locale]/calculator/zlomky-new/page.tsx
@@ -1,8 +1,0 @@
-// src/app/[locale]/calculator/zlomky-new/page.tsx
-'use client';
-
-import FractionsCalculator from '@/components/calculators/FractionsCalculator';
-
-export default function FractionsTestPage() {
-  return <FractionsCalculator />;
-}
